Revising how I edit a questionnaire (see description)
The easier way is to save each part of the questionnaire, starting from
the innermost elements.
Foreach sentence, I check if it already exists and if so, I update the
fields (one by one), otherwise I create a new one. Then the same thing
for questions, answers, sections and eventually the questionnaire.
In this way EF shouldn't complain about reference integrity.

Also, I may think about bring the API part, into the database project.
Could add just the API support without all the mess of files that the
WebAPI project model brings with it. In this way I won't need
"Repository" classes that for now are just a waste of time and
duplicated code.