Changing the way the repository works (read description)
To delete an entity it should be marked as deleted. New partial classes
of every entity implement a Deleted boolean that indicates if the entity
should be deleted. The UpdateEntitiesState method in the contex, will
look for this flag to mark the entity to be deleted to the next
SaveChanges.
This changes were applied ONLY to the SECTION entity. Should be
replicated to all the other entities.
I may think about changing the EF to code first...